refactor: remove dead code and update docs
