+++ src/components/ui/LoadingState/index.ts
@@ -1,8 +1,6 @@
 export { 
   LoadingState, 
-  LoadingSpinner, 
   LoadingSkeleton, 
-  LoadingDots, 
   LoadingPulse,
   type LoadingVariant,
   type LoadingSize
